+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+<a name="1.0.6"></a>
+# 1.0.6 (2018-10-24)
+
+## Feature
+- **fixedColumnLeft:** prevent default 0, Using with fixed left columns.
+- **fixedColumnRight:** prevent default 0, Using with fixed right columns.
+  
 <a name="1.0.5"></a>
 # 1.0.5 (2018-10-24)